Cyber Crime Unit Online Complain
With CRS (Crime Reporting System), public and police are both benefitted as it reduces the time and increases the chance of solving the case. Which then reduces the Crime rate.  Public can login to the site and then register their complaint, user friendly page guides even a layman with basic English knowledge to register their complaints.
  Duplicate records are eliminated and forging with dummy complaints are strongly controlled with the help of One Time Password (OTP) which will be sent to the public who registers with us. Only verified users record will be saved, else the complaint will be discarded.
If the Public has any evidence with respect to the crime happened, they can attach the evidence in the type of any image file and save it with their mobile number and complaint type so that it is easy for police officials to solve the case.
Once Mobile Number is verified, the complaint is registered in the server.
Police Officials can view the complaints with respect to there are by logging with their station name and password they registered with the system.  Interactive UI enables the users to select their area station to book their complaints, thereby reducing time.
Officials will there logged into their system, and can check any new complaints for their station received. If so they can check for the complaint and  mark the complaint as Ongoing once the complaint is filed, the civilian who had made the complaint can log into the system with their mobile number, email id and the OTP received to their mobile number at the time of complaint , users can also request for OTP if they have forgotten through mail or sms.
They then check there complaint status and can follow up the case status through this system.
Officials can also check for the local criminals using our powerful searching system, and can track the criminal’s current record and then update.
For every single user (police (or) civilians), separate session id is maintained and no one can access their account without a valid credentials.